+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unrelease 1.1.3
+### Changed
+- **ProductCardFeatures**
+    - devide showcaseParams prop into firstParam and secondParam
+- **package.json**
+    - add script jest updateSnapshot
 ## Release 1.1.2
 ### Changed
 - **icons/Basic/24/Hide_24.svg**